chore(clippy): fix --all-features --all-targets clippy lints
Pre-existing lint debt in cera core (simd/wgpu/cpu/turboquant) that only
fires under --all-features + --all-targets (not the default CI gate). No
runtime behavior change: needless_range_loop rewritten to iterators where a
clean single-slice case / #[allow] with rationale on kernel loops indexing
multiple buffers; div_ceil; #[allow(dead_code)] on cfg-conditional SIMD gemm
fns with call-site notes; #[allow(too_many_arguments)]; doc-list indentation.

+++ cera/tests/detok_metal_parity.rs
@@ -229,7 +229,7 @@ fn multi_frame_stability() {
     // 8 frames = 48 tokens → wraps the SWA window (30) 1.6 times
     let all_codes: Vec<[i32; 8]> = (0..8)
         .map(|i| {
-            let base = (i * 100 + 50) as i32;
+            let base = i * 100 + 50;
             [
                 base,
                 base + 100,
